requirements.txt
google-generativeai
requests
python-dotenv
Jinja2
premailer
feedparser
